refactor: closes #15 Update PillList and TabList components
Update PillList component to use an unordered list (ul) instead of a div for better semantic structure. Each pill item is now wrapped in a list item (li) with appropriate schema.org attributes.

Update TabList component to add horizontal scrolling for tabs when there are too many to fit in the container. The overflowMobile prop is introduced to control whether the scrolling behavior should be enabled on mobile devices.

These changes improve the structure and user experience of the PillList and TabList components.